disk: Add metadata store insertion for root block
Add insertion of root block for the metadata tree. This seems to have
a problem copying the contents of on disk mappings, it appears that in
some instances the backing map does not support reads.

+++ src/main/java/edu/rit/cs/disk/MetadataExtent.java
@@ -24,7 +24,7 @@ class MetadataExtent
 	 * @param buf Mapped buffer region.
 	 * @param exn Extent number on disk
 	 */
-	public MetadataExtent(ByteBuffer mbuf, int exn)
+	public MetadataExtent(ByteBuffer mbuf, long exn)
 	{
 		this.mbuf = mbuf;
 		this.exn = exn;
